requirements.txt
Django==2.2.7
django-rest-multiple-models==2.1.3
djangorestframework==3.10.3
djangorestframework-csv==2.1.0
psycopg2-binary==2.8.4
pytz==2019.3
rpdb==0.1.6
six==1.13.0
sqlparse==0.3.0
unicodecsv==0.14.1
uWSGI==2.0.18
